Added update for traffic light

+++ src/trafficlight.h
@@ -8,8 +8,8 @@ class TrafficLight {
 private:
     std::pair<float, float> position;        // position of traffic light
     bool lightState = false;                 // false = red, true = green
-    double timer = 0.0;                      // timer for light state
-    double redTimer = 0.0, greenTimer = 0.0; // timers for red and green lights
+    float timer = 0.0f;                      // timer for light state
+    float redTimer = 0.0f, greenTimer = 0.0f; // timers for red and green lights
     Texture2D *red;
     Texture2D *yellow;
     Texture2D *green;
